# [SPARK-25379][SQL] Improve AttributeSet and ColumnPruning performance