@maropu 2018-09-14
Can we replace the syntax(`(ouputSetA -- outputSetB).nonEmpty`)  in other places, too? e.g., 
https://github.com/apache/spark/blob/9deddbb13edebfefb3fd03f063679ed12e73c575/sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/optimizer/CostBasedJoinReorder.scala#L294