@mgaido91 2018-09-14
@maropu yes, that can be done as well, but I think the main focus of this PR is the `ColumnPruning` rule, so I think it would be great to do that in a separate PR. What do you think? Thanks.