@maropu 2018-09-14
IIUC this pr targets to improve `AttributeSet` operations, so all the places get the same luck with `ColumnPruning`? If so, I think its ok to fix all the places in this pr. cc: @gatorsmile