@mgaido91 2018-09-14
@maropu anyway I checked and that is the only other places where this pattern happens. So I am ok including it here. The point is that there the situation is a bit different, ie. it is not an `(AttributeSet -- AttributeSet).nonEmpty` case, but it is a `(AttributeSet -- Seq[Attribute]).nonEmpty`. So I am not sure the performance gain/difference will be significant in this case.

I'll try and do some benchmarks and I'll do the change if I see a significant difference. Thanks.