@maropu 2018-09-18
oh, yea, thanks! I wrongly mixed up `(AttributeSet -- Seq[Attribute]).nonEmpty` with this case.