@mgaido91 2018-09-24
I also run on the TPCDS and TPCH benchmark with 10 runs:

Rule | Effective After | Effective Before | Total After | Total Before | % Eff | % Total
-- | -- | -- | -- | -- | -- | --
org.apache.spark.sql.catalyst.optimizer.ColumnPruning (TPCDS) | 8492237639 | 9927405142 | 37554775945 | 44771729889 | 85.5433773 | 83.8805560
org.apache.spark.sql.catalyst.optimizer.ColumnPruning (TPCH) | 214235083 | 292912646 | 933477165 | 1166285066 | 73.1395813 | 80.0385079
-- | -- | -- | -- | -- | -- | --

Which confirm a ~20% improvement after the change (with wide-column datasets, the improvement is much higher as showed by the benchmark referenced in the PR description).